Features: (1) Made it so that the scene has sole ownership of all the game objects. The only thing game objects own are their components for now, they don't own their children, they just have accesss to them. (2) You can now remove game objects from the scene and have the option to keep or remove the children of that object (if any) from the scene as well. If you don't remove them then the parent object(if any) will get the children.
Know Issue: Right now if you DON'T add a game object to the scene it will create a memory leak. Probably gonna use RAII to fix it and wrap the creation of a game object around a function of some sort.

GUI rendering is not very good right now, started working on a GUI Manager which will handle all the necessary GUI for the engine.