Sound System Progress || Logger Changes
Features: (1) Made some progress with the sound system. (2) Correctly linked SDL mixer and you can now load an play sounds. (3) Audio file paths are now going to be stored in a map with their name as an enum for the key. [Need to find a way to make this enum editable for the users of the engine, it will stay like this for now].

Changes:
(1) Changed how the logger works so that it can correctly show you which line a call was given from.