Audio Partially working
The functionality is there. Instead of using the multithreading aproach which I couldn't manage to get working, I will take advantage of the service locator and make another AudioSystem that is specific to the web that doesn't use multithreading.

To avoid loading in the hot path, for now, I will probably make all sounds be pre-loaded in the web build.